docs: HALO credentials design. (#1380)

+++ toolchains/ridoculous/src/processor.ts
@@ -2,20 +2,18 @@
 // Copyright 2022 DXOS.org
 //
 
-import debug from 'debug';
+import chalk from 'chalk';
 import * as fs from 'fs';
 import glob from 'glob';
 import * as path from 'path';
-import * as process from 'process';
 import { read } from 'to-vfile';
 
 import { createParser } from './parser.js';
 
-const log = debug('dxos:ridoculous:processor');
-
 interface Options {
   autoNumber?: boolean
   baseDir?: string
+  dryRun?: boolean
   files?: string
   html?: boolean
   outDir?: string
@@ -24,30 +22,36 @@ interface Options {
 
 export const processFiles = async ({
   autoNumber,
-  baseDir = process.cwd(),
-  files = 'docs/**/*.md',
-  html = false,
-  outDir = './out',
+  baseDir,
+  dryRun,
+  files,
+  html,
+  outDir,
   verbose
 }: Options = {}) => {
-  const parser = createParser({ autoNumber, baseDir, html });
+  const parser = createParser({ autoNumber, baseDir, html, verbose });
 
   const globFiles = glob.sync(files);
   for (const filename of globFiles) {
-    log(`Parsing: ${filename}`);
-
-    // TODO(burdon): Catch errors.
-    const text = await parser.process(await read(filename));
-
-    const parts = path.parse(filename);
-    const f = path.format({ ...parts, base: undefined, ext: html ? '.html' : '.md' });
-    const outFilename = path.join(outDir, path.relative(baseDir, f));
-    const dirname = path.dirname(outFilename);
-    if (!fs.existsSync(dirname)) {
-      fs.mkdirSync(dirname, { recursive: true });
+    if (dryRun || verbose) {
+      console.log(`Parsing: ${chalk.green(filename)}`);
     }
 
-    fs.writeFileSync(outFilename, text.toString() + '\n', 'utf8');
-    console.log(`Updated: ${outFilename}`);
+    // TODO(burdon): Catch errors.
+    // https://github.com/vfile/vfile#filehistory
+    const text = await parser.process(await read(filename));
+
+    if (!dryRun) {
+      const parts = path.parse(filename);
+      const f = path.format({ ...parts, base: undefined, ext: html ? '.html' : '.md' });
+      const outFilename = path.join(outDir, path.relative(baseDir ?? '.', f));
+      const dirname = path.dirname(outFilename);
+      if (!fs.existsSync(dirname)) {
+        fs.mkdirSync(dirname, { recursive: true });
+      }
+
+      fs.writeFileSync(outFilename, text.toString() + '\n', 'utf8');
+      console.log(`Updated: ${chalk.green(outFilename)}`);
+    }
   }
 };
